Feat: add formatting to the calculated result.
This feature ensures that the calculated result follows the behaviour of the iphone calculator. All commas, decimals, rounded numbers, and and exponential results appear to be in-line with the iphone at the time of writing the code. This code also ensures that the calculated result will always stay inside of the phone display